checking

+++ src/app/api/posts/myneed/[id]/route.js
@@ -1,7 +1,16 @@
 import { NextResponse } from "next/server";
+import { getAuthSession } from "../../../../../utils/auth";
 
 export const DELETE = async (req, { params }) => {
   const { id } = params;
+
+  const session = await getAuthSession();
+
+  if (!session) {
+    return new NextResponse(
+      JSON.stringify({ message: "Not Authenticated!" }, { status: 401 })
+    );
+  }
 
   try {
     await prisma.post.delete({
@@ -9,6 +18,7 @@ export const DELETE = async (req, { params }) => {
         id: id,
       },
     });
+    return new NextResponse(204);
   } catch (error) {
     console.log(error);
   }
